--- alternate_simulation_inputs/cool_down.in
@@ -1,42 +1,0 @@
-Simulated annealing, 20 ps; no velocities read in; Langevin dynamics thermostat
-
- &cntrl
-    imin=0,
-    ntx=1,
-    irest=0,
-    ntxo=2,
-    ntpr=200, 
-    ntwx=200, 
-    ioutfm=1,
-    nstlim=20000, 
-    dt=0.001,
-    ntt=3,
-    gamma_ln=2,
-    ig=-1,
-    ntc=2,
-    ntf=2,
-    ntb=0,
-    cut=15.0, 
-    pencut=-0.001, 
-    vlimit=10, 
-    nmropt=1
- /
- &ewald
-    eedmeth=5,
- /
-#
-#from steps 0-18000: maintain temp at USER_TEMP K
-#from steps 18001-20000: final cooling to 0 K
-#
- &wt type='TEMP0',istep1=0,istep2=18000,value1=USER_TEMP,value2=USER_TEMP, /
- &wt type='TEMP0',istep1=18001,istep2=20000,value1=USER_TEMP,value2=0.0, /
-#
-#from steps     0- 3000: slowly ramp up restraint strengths
-#from steps  3001-20000: maintain restraints at full strength
-#
- &wt type='REST', istep1=0,istep2=3000,value1=0.1,value2=1.0, /
- &wt type='REST', istep1=3001,istep2=20000,value1=1.0,value2=1.0, /
-
- &wt type='END'  /
-LISTOUT=POUT
-DISANG=RST
