--- alternate_simulation_inputs/md.in
@@ -1,30 +1,0 @@
-normal MD, 20 ps; no velocities read in; Langevin dynamics thermostat
-
- &cntrl
-    imin=0,
-    ntx=1,
-    irest=0,
-    ntxo=2,
-    ntpr=200, 
-    ntwx=200, 
-    ioutfm=1,
-    nstlim=20000, 
-    dt=0.001,
-    ntt=3,
-    temp0=USER_TEMP,
-    gamma_ln=2,
-    ig=-1,
-    ntc=2,
-    ntf=2,
-    ntb=0,
-    cut=15.0, 
-    pencut=-0.001, 
-    vlimit=10, 
-    nmropt=1
- /
- &ewald
-    eedmeth=5,
- /
- &wt type='END'  /
-LISTOUT=POUT
-DISANG=RST
