+++ fold_protein.json
@@ -1,13 +1,14 @@
 {
 	"name": "1ubq",
-	"fasta_file_path": "~/Scripts/git/OpenFold-amber/Test_Suite/1UBQ_example/1UBQ.fasta",
-	"distance_restraints_file_path": "~/Scripts/git/OpenFold-amber/Test_Suite/1UBQ_example/1UBQ_8col.dist",
-	"torsion_restraints_file_path": "~/Scripts/git/OpenFold-amber/Test_Suite/1UBQ_example/1UBQ_5col.angl",
-	"simulated_annealing_input_file_path": "~/Scripts/git/OpenFold-amber/siman.in",
-	"tordef_file_path": "~/Scripts/git/OpenFold-amber/tordef.lib",
+	"fasta_file_path": "../Test_Suite/1UBQ_example/1UBQ.fasta",
+	"distance_restraints_file_path": "../Test_Suite/1UBQ_example/1UBQ_8col.dist",
+	"torsion_restraints_file_path": "../Test_Suite/1UBQ_example/1UBQ_5col.angl",
+	"simulated_annealing_input_file_path": "../siman.in",
+	"tordef_file_path": "../tordef.lib",
 	"forcefield": "leaprc.protein.ff14SB",
 	"distance_force_constants": [70.0],
 	"torsion_force_constants": [40.0],
 	"temperatures": [600.0],
-    	"annealing_runs": 1
+  "annealing_runs": 1,
+  "max_threads": 3
 }
