--- temp_codes/200_steps_siman.in
@@ -1,58 +1,0 @@
-Simulated annealing, 20 ps; no velocities read in; Langevin dynamics thermostat
-
- &cntrl
-    imin=0,
-    ntx=1,
-    irest=0,
-    ntxo=2,
-    ntpr=200, 
-    ntwx=200, 
-    ioutfm=1,
-    nstlim=200, 
-    dt=0.001,
-    ntt=3,
-    tempi=100,
-    gamma_ln=2,
-    ig=-1,
-    ntc=2,
-    ntf=2,
-    ntb=0,
-    cut=15.0, 
-    pencut=-0.001, 
-    vlimit=10, 
-    nmropt=1
- /
- &ewald
-    eedmeth=5,
- /
-#
-#Simple simulated annealing algorithm:
-#
-#from steps 0 to 50: heat the system to USER_TEMP K
-#from steps 51-180: re-cool to 100 K with long tautp
-#from steps 181-200: final cooling to 0 K with short tautp
-#
- &wt type='TEMP0', istep1=0,istep2=50,value1=USER_TEMP,
-            value2=USER_TEMP,    /
- &wt type='TEMP0', istep1=51, istep2=180, value1=USER_TEMP,
-            value2=100.0,     /
- &wt type='TEMP0', istep1=181, istep2=200, value1=0.0,
-            value2=0.0,     /
-
- &wt type='TAUTP', istep1=0,istep2=50,value1=0.4,
-            value2=0.4,     /
- &wt type='TAUTP', istep1=51,istep2=180,value1=4.0,
-            value2=4.0,     /
- &wt type='TAUTP', istep1=181,istep2=190,value1=1.0,
-            value2=1.0,     /
- &wt type='TAUTP', istep1=191,istep2=200,value1=0.1,
-            value2=0.05,    /
-
- &wt type='REST', istep1=0,istep2=30,value1=0.1,
-            value2=1.0,  /
- &wt type='REST', istep1=31,istep2=200,value1=1.0,
-            value2=1.0,  /
-
- &wt type='END'  /
-LISTOUT=POUT
-DISANG=RST
